refactored pad calibration, lots of optimizations, lcd update refactoring, misc

+++ src/Firmware/interface/encoders/Encoders.h
@@ -13,7 +13,7 @@ class Encoders
 
     private:
     void handleEncoder(uint8_t encoderNumber, bool direction, uint8_t steps);
-    uint32_t lastStepTime[MAX_NUMBER_OF_ENCODERS];
+    uint8_t lastStepTime[MAX_NUMBER_OF_ENCODERS];
 };
 
 extern Encoders encoders;
